requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.35.14
botocore==1.35.14
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cohere==5.9.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
deepeval==1.1.6
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
dotenv-python==0.0.1
enum34==1.1.10
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
faiss-cpu==1.8.0.post1
fastavro==1.9.7
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.11.0
grpcio==1.63.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.2.16
langchain-cohere==0.2.4
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.65
langchain-groq==0.1.9
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langcodes==3.4.0
langsmith==0.1.116
language_data==1.2.0
llama-cloud==0.0.17
llama-index==0.11.7
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.0
llama-index-core==0.11.7
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.44.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.1
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
pypdf==4.3.1
pysbd==0.3.4
pytest==8.3.2
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.1.16
rank-bm25==0.2.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentry-sdk==2.13.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
srsly==2.4.8
stack-data==0.6.3
striprtf==0.0.26
sympy==1.13.2
tabulate==0.9.0
tenacity==8.4.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.5
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
widgetsnbextension==4.0.13
wrapt==1.16.0
nbformat==5.10.4
xxhash==3.5.0
yarl==1.10.0
zipp==3.20.1
